Up Code Motion pass
Summary:
This introduces a new pass that eliminates gotos by moving trivial const-instructions before a conditional branch if that doesn't interfere with anything, and then allows to remove the conditional branch target block.

For example:

       IF_EQZ v2, L1
       CONST v0, 1
       ... (GOTO elsewhere or RETURN or THROW)
   L1: CONST v0, 0 // where L1 is only reachable via the above IF-instruction
       GOTO L2

becomes

       CONST v0, 0
       IF_EQZ v2, L2
       CONST v0, 1
       ...

The savings are modest, but directly translate to compressed savings, as gotos are notoriously uncompressible.
And I no longer have to look at such avoidable gotos while reading dex code...

Most of the time when this transformation applies, exactly one const instruction is moved. If the two outgoing branches are called equally often, then the total number of instructions executed doesn't change (in one case, an extra const instruction is executed, in the other case we save a goto). So this should be rather perf-neutral in terms of execution instructions count, and a size win by removing gotos from the code. Smaller code means less I/O, so maybe even faster.

Reviewed By: int3

Differential Revision: D14393038

fbshipit-source-id: 5db1d65858a0e02360c659fdc2574adbd9e412c8