Make CrossDexRefMinimizer::worst deterministic
Summary: At the beginning of each dex, a "worst" class is picked --- the class with the most dependencies. However, this was done in a way that didn't distinguish between classes with equal "badness". This diff introduces a tie-breaker, deterministically prefering the class that was inserted earlier (a.k.a. appears earlier in the scope list).

Reviewed By: emma0303

Differential Revision: D14415274

fbshipit-source-id: fe2039b05f69d56ced9bc9b97553e1060c43f6ef