Add support to parse APK of Dex version 37.
Summary:
This patch adds an option to Redex to support parsing APK of Dex version 37 with Java 8 features. When this is enabled, Redex can parse both Dex version 35 and version 37, and the output's Dex version is the same as input.
Some opt passes (RemoveInterfacePass) are known to crash with Java 8 features, and more supports would be needed in the future.

Reviewed By: thezhangwei

Differential Revision: D14379214

fbshipit-source-id: d33beac8afbd76f29241982b2b70db6d6b2d64ef