Emit list of GraphQL queries that were removed by RMU and others
Summary:
GraphServices generates metadata for queries in an app. This is a binary file `assets/graph_metadata.bin` that lives in the APK, gets decompressed and read at runtime.

Because the dep graph of fb4a is, shall we say untidy, it is suspected that this binary file might be containing metadata for queries that are provably dead via RemoveUnreachable, or meant for other apps like Messenger, Workplace, etc.

The idea here is that we could pretty easily output a list of such classes; at release build time after Redex runs we could ask the GraphQL build tools to rebuild the file based on this list of dead classes to regenerate the (hopefully smaller) binary file.

Reviewed By: thezhangwei

Differential Revision: D14694032

fbshipit-source-id: a9dd54bf1d94ad6f477c9a629dcfea079a1cc9eb