Better support for ctor type tag param alternative locations
Summary:
There's a new case in Type Erasure where the input program handles the type tag, but we have to construct a dispatch using the existing type tag value. More specifically for the ctor dispatches, we need to be able to find out which param passed in is the type tag and switch on it.

I made a couple of changes to make this work so far:
- Add an optional `type_tag_param_idx` in `dispatch::Spec` to encode the logical index of the type tag param
- Scan the type list of the ctor arg list and try to infer the position of the type tag based on it's type
- Add an instr test hierarchy to verify the change

Reviewed By: Feng23

Differential Revision: D14634408

fbshipit-source-id: 9511008f69f584b72d8fbec61d13109304e50337