replace assert with new name redex_assert
Summary:
We have a couple of places in redex call the assert function in standard C library. Most of them never intended to do that but due to naming collision between the `assert.h` and our implementation of assert, they are bound to the standard library.

Now since we want to use `RedexException`, it's increasing important that we use it consistently in redex. In order to fix this, I am doing this codemod.

Reviewed By: int3

Differential Revision: D14781739

fbshipit-source-id: 3d05530caf9b5bbcbf0b1f8684b9a5212b479d08