Cross-dex-ref minimization: Consider global frequencies
Summary:
Certain kinds of references should not be considered in the class-selection prioritization algorithm that tries to minimize cross-dex references.

Those are references that...
- only ever appear in a single class, and
- are so common they most likely will get referenced from all dexes anyway, e.g. Ljava/lang/Object;

The result is an increased efficiency of the cross-dex ref minimization algorithm --- if you consider all the diffs in this stack; this diff lays some groundwork.

Reviewed By: minjang

Differential Revision: D13777329

fbshipit-source-id: 0fd8628dbc6cb38b77dd31e5cc8c38c82118032e