Sink super invocations from switch statements to return block in merger constructors
Summary:
This patch tries to reduce 'super' invocations in merger constructors.
In merger constructors, each type calls a 'super' constructor in their switch statements. If the super constructors are the same, we could sink them to a common invocation in the return block followed by each switch statements. The patch erases 'super' invocations in switch statements, moves their arguments to the same registers, and inserts a common invocation in return block. If the 'move' are redundant, they should be optimized by other passes.

Reviewed By: thezhangwei

Differential Revision: D14795514

fbshipit-source-id: e4395340b7995c4063c22413844f4a20f46342f1