Change ModelMethodMerger::sink_common_ctor_to_return_block to use non-editable CFG
Summary: IRCode->insert_before is not expected to work after IRCode->build_cfg(/* editable */ true) because the newly built CFG steals the m_ir_list containing the instructions. But in practice, the scenario works because instructions can link to new instructions that are not in the list. However, this is flaky. The patch changes ModelMethodMerger::sink_common_ctor_to_return_block to use non-editable CFG instead, and will update to editable CFG once remaining pieces of editable CFG are checked in.

Reviewed By: justinjhendrick

Differential Revision: D14895473

fbshipit-source-id: ba28074f25e71d0cea8a84c5cf7faf6b42663105